Refactor scheduling logic and enhance domain models for improved functionality
- Updated the `remaining_hours` and `hours_exhausted` functions in the `assignment.py` file to support additional parameters for `schedule_kind`, allowing for more flexible hour calculations based on different scheduling types.
- Introduced new constants and utility functions from `schedule_variant.py` to streamline the handling of monthly and temporary scheduling scenarios.
- Enhanced the `ScheduleCell` model by adding `schedule_kind` and `week_index` fields to better manage scheduling variations and improve data integrity.
- Modified the `TeachingAssignment` model to change `hours_per_week` from `Integer` to `Float`, accommodating fractional hour values for more accurate workload representation.
- Updated various service files to incorporate the new scheduling logic, ensuring that functions like `placed_counts`, `remaining_for`, and others correctly utilize the new parameters for improved accuracy in scheduling calculations.

These changes collectively enhance the scheduling system's flexibility and accuracy, providing a more robust framework for managing teaching assignments and classroom schedules.

+++ app/models/teaching_assignment.py
@@ -1,5 +1,5 @@
 """Teaching assignment model."""
-from sqlalchemy import Column, ForeignKey, Integer
+from sqlalchemy import Column, Float, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
 from app.db import Base
@@ -13,7 +13,7 @@ class TeachingAssignment(Base):
     subject_id = Column(Integer, ForeignKey("subjects.id"), nullable=False)
     teacher_id = Column(Integer, ForeignKey("teachers.id"), nullable=True)
     class_id = Column(Integer, ForeignKey("school_classes.id"), nullable=False)
-    hours_per_week = Column(Integer, nullable=False)
+    hours_per_week = Column(Float, nullable=False)
     group_number = Column(Integer, nullable=True)
     preferred_classroom_id = Column(Integer, ForeignKey("classrooms.id"), nullable=True)
