+++ app/services/assignment/types.py
@@ -16,7 +16,7 @@ class SubjectAssignClassData:
     id: int
     name: str
     grade: int
-    hours_per_week: int
+    hours_per_week: float
     teacher_ids: list[int]
     is_split: bool
